// Project_Apex/frontend/src/api/models/KycApplicationPublic.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type KycApplicationPublic = {
    id: string;
    user_id: string;
    email: string;
    kyc_status: string;
    kyc_submitted_at: string | null;
    legal_first_name: string | null;
    legal_last_name: string | null;
    date_of_birth: string | null;
    phone_number: string | null;
    country: string | null;
    risk_assessment_score: number;
};